Setting MelModel per game - override parent record elements: RRR
IIRC all this started cause Sharlikran needed to override the
MelModel for SSE but oddly the Oldrim definition was used. Turns out
was defined in brec and imported and used in skyrim/records which was
imported first thing in sse records. So we needed a mechanism for
allowing overrides of record elements of parent games from derived
games. This is pretty much the limit of the current API. Current
solution (after failed attempts like using `__all__`) imposes two
conditions: child games init must import from package records.py first.
And we need unused imports in record files of child games as they are
accessed from there. The latter one was unmasked because we got rid
of the star imports which was a major todo in wip game handling
design. At least now the complexity is on the surface. Because star
imports were gone we needed to import explicitly from other modules
(another bonus). Took the opportunity to clean up and added comments
on seemingly unneeded imports.
In the process turned out MelMODS (used in MelModel) was functionally
identical in various games so moved to brec - ignore the dust.

This finalizes record modules import system, by removing the star
imports from records files. It seems the most satisfactory solution
as more explicit. We explicitly import the records we need from
parent game. Only caveat is that the import order of parent vs child
game record files matters, but that is just a manifestation of the
inherent complexity in game hierarchy's handling. What remains is
importing .constants.py files only as needed for loaded game - see # 358

Sharlikran contributed the MelModel for sse and fo4 (the skyrim he
edited afterwards so it got identical to the sse so the initial need for
the MelModel hack was no more, but the override seems needed for fo4).
Infernio split some unrelated changes (see next commit)
We needed 8 months for the dust to settle and have only the needed
changes in here. Some scraps from squashed commits and diffs follow.

Sharlikran:

Updated class MelModel

This discards textureHashes since they are not used in loose files.
Which fixes an issue with adding Form Version 44 to the header of
Skyrim SE plugins that contain textureHashes.

----------------------- Mopy/bash/game/skyrim/records.py -----------------------
index 88219f9b..bac0256c 100644
@@ -54,2 +54,13 @@ class _MelModel(brec.MelGroup):

+        class MelModelHash(MelBase):
+            """textureHashes are not used for loose files. There is never a
+            Bashed Patch, 0.bsa. The record will be read if present but no
+            defaults are set and the record will not be written."""
+            def loadData(self, record, ins, sub_type, size_, readId):
+                MelBase.loadData(self, record, ins, sub_type, size_, readId)
+            def getSlotsUsed(self):
+                return ()
+            def setDefault(self, record): return
+            def dumpData(self, record, out): return
+
         def __init__(self, attr='model', subType='MODL'):
@@ -58,3 +69,3 @@ def __init__(self, attr='model', subType='MODL'):
             MelGroup.__init__(self, attr, MelString(types[0], 'modPath'),
-                              MelBase(types[1], 'modt_p'),
+                              self.MelModelHash(types[1], 'textureHashes'),
                               MelMODS(types[2], 'alternateTextures'), )

[FO4] Updated class MelModel

Note diffs from skyrim one:

---------------------- Mopy/bash/game/fallout4/records.py ----------------------
index c5148a12..bf98e666 100644
@@ -84,8 +84,9 @@ class _MelModel(MelGroup):
         # MODB and MODD are no longer used by TES5Edit
-        typeSets = {'MODL': ('MODL', 'MODT', 'MODS'),
-                    'MOD2': ('MOD2', 'MO2T', 'MO2S'),
-                    'MOD3': ('MOD3', 'MO3T', 'MO3S'),
-                    'MOD4': ('MOD4', 'MO4T', 'MO4S'),
-                    'MOD5': ('MOD5', 'MO5T', 'MO5S'),
-                    'DMDL': ('DMDL', 'DMDT', 'DMDS'), }
+        typeSets = {'MODL': ('MODL', 'MODT', 'MODC', 'MODS', 'MODF'),
+                    'MOD2': ('MOD2', 'MODT', 'MO2C', 'MO2S', 'MO2F'),
+                    'MOD3': ('MOD3', 'MODT', 'MO3C', 'MO3S', 'MO3F'),
+                    'MOD4': ('MOD4', 'MODT', 'MO4C', 'MO4S', 'MO4F'),
+                    'MOD5': ('MOD5', 'MODT', 'MO5C', 'MO5S', 'MO5F'),
+                    # Destructible
+                    'DMDL': ('DMDL', 'DMDT', 'DMDC', 'DMDS'), }

@@ -105,5 +106,9 @@ def __init__(self, attr='model', subType='MODL'):
             types = self.__class__.typeSets[subType]
-            MelGroup.__init__(self, attr, MelString(types[0],'modPath'),
-                              self.MelModelHash(types[1],'textureHashes'),
-                              MelMODS(types[2],'alternateTextures'), )
+            MelGroup.__init__(
+                self, attr, MelString(types[0], 'modPath'),
+                self.MelModelHash(types[1], 'textureHashes'),
+                MelOptStruct(types[2], 'f', 'colorRemappingIndex'),
+                MelOptStruct(types[3], 'I', (FID,'materialSwap')),
+                MelBase(types[3], 'modf_p'),
+                )

Utumno:

Move MelMODS to brec. Were identical apart from:

@@ -500,2 +500,3 @@ def loadData(self, record, ins, sub_type, size_, readId):
         insUnpack = ins.unpack
+        insRead32 = ins.readString32
         count, = insUnpack('I',4,readId)
@@ -504,5 +505,5 @@ def loadData(self, record, ins, sub_type, size_, readId):
         for x in xrange(count):
-            string = ins.readString32(readId)
+            string = insRead32(readId)
             fid = ins.unpackRef()
-            index, = ins.unpack('I',4,readId)
+            index, = insUnpack('I',4,readId)
             dataAppend((string,fid,index))

I kept the inlines

Under # 358

Co-authored-by: Sharlikran <danielhmpdx16@gmail.com>
Co-authored-by: MrD <the.ubik@gmail.com>