Hack for inconsistent bsa/b2a path separators: EEE ask bsas are backslashes only??
See discussion in:
https://afkmods.iguanadons.net/index.php?/topic/4966-wrye-bash-all-games/&do=findComment&comment=170410
https://afkmods.iguanadons.net/index.php?/topic/4966-wrye-bash-all-games/&do=findComment&comment=170367

Hack as it's performance intensive, but introducing heuristics won't
work as some ba2 files have _both_ filename separators. Gone down that
path as Path has backslashes normalized so in has_assets we only need
to lower. Would it make sense to only expect backslashes (so lower())
for bsas and normcase() for ba2s?