[TES5, FO4] WIP: Fixed leveled list patcher discarding 0s
WIP simply because this needs more testing, which I'll do tomorrow.

To see this in action, try building a BP in Skyrim (wihout this commit)
with the 'Scarcity' mod installed. It changes a bunch of LVLD entries
to be 0. Unfortunately, the patcher checked to see if a value was
present via an implicit bool() conversion - and bool(0) == False.
Therefore, none of those LVLD changes got carried over, since WB
thought Scarcity just didn't have any values present there.

This changes MreLeveledListBase.mergeWith to use an explicit comparison
with None instead to determine if a value is present. Should only affect
Skyrim (LE & SE) and Fallout 4, since the other games don't use
MreLeveledListBase.

Also seems to have fixed an issue where the BP would sometimes not carry
forward changes to the flags of an LVLI record - probably happened
because flags are (internally) treated as integers, and so a subrecord
with no flags checked counts as 0 as well.