Drop MelModelHash in favor of MelNull
MelNull already does what we want - the definitions are almost
identical! The only real difference is that MelModelHash also saves the
texture hashes as an attribute of the record. This is obviously totally
unnecessary - we don't want to do anything with those hashes.

To that end, let's just drop those classes and use MelNull.