Be less cautious in Path.version
Drops a way too broad except clause there. The only reasonable
exception I could see coming from get_file_version is a ValueError from
the new string code, so I added an except there to make sure.

If this does end up breaking something, it points to a way deeper issue
that actually needs investigating, not silencing.

Fix crash on startup due to xSE version check

Turns out that there was a reason for the insanely broad except clause
in Path.version. StatusBar_Button.obseVersion did not check if the xSE
file actually exists before checking its version - which obviously
errors.

Obviously, the correct solution here is not to silence that erorr, but
to make obseVersion check if the file actually exists before calling
strippedVersion on it.

Add support for reading versions of .lnk files

This fixes a crash when any of those are added to the 'Apps' folder and
also shows their versions when the 'app version' status bar feature is
enabled - so we even came out of this with a new feature!

Also added some minor try-except clauses around the win32 api calls
there now that I understand what actually causes them.