WIP: Use FileInfo.get_masters() where possible
Note: This relies on the cosaves rework, specifically the 'cosaves.py -
beta' commit and the 'correct save masters' commit.

Much easier to find the definition and search usages for than the
previous '.header.masters' code, which is now only used where necessary.
Will also allow us to be much more flexible in the future if masters
become more complex than that.

Obviously touches a ton of places, so WIP, but it shouldn't have any
functional impacts.