WIP: Show ESP/M and ESL counts in the bottom right corner
Only on games with ESL support, obviously. Had to expand the bottom
right part of the status bar for ESL games to do this, and took that
opportunity to factor out some common code in the BashStatusBar methods.

As the TODO mentions, I'm not really happy with this. It feels like we
should already be storing this information somewhere instead of having
to compute it on the fly. I've probably just overlooked the spot where
it's stored though.

Improve performance of ESP/M & ESL counter

Use a reduce with lambda instead of iterating over the entire list.
Should be much faster, since we didn't need the full lists to begin
with.