Convert to long FormIDs before patching Stats
This _properly_ fixes the crashes and the ITMs caused by the 'Critical
Effect' field of the CRDT subrecord of the WEAP record.

The cause was that the record definitions define criticalEffect to be a
FID, but to properly initialize FormIDs, convertToLongFids() must first
be called. Otherwise, they remain as integers. Later on in the patch
building process, we make a copy of each record from the source mods, to
be included in the final BP, but this copy procedure
(MreRecord.getTypeCopy, if you're curious) converts FormIDs when
copying. This resulted in us having cached mod data where criticalEffect
was an int, and a copy of that data where criticalEffect was a FormID.

Long FormIDs are tuples of a string and an integer, while short FormIDs
are just integers. Obviously, a comparison between the two will show
them as different - causing the BP to think that _every_ record that
occurs in a source mod had changes that needed to be preserved - hence
the ITMs.

I haven't figured out how this was responsible for the crashes, but I
can confirm that they're fixed as of this commit.

Supersedes 'Fix Duplicate Records In Stats Parsing' (73be14f51d04aefbecc07b92cb5c6af52febf5e8).