Skip BAIN refresh for xEdit caches
For the same reasons we skip refreshes for xEdit backups, we should also
skip xEdit caches. Also edited the README to match.

Turns out that SSE never received the edits to skip the 'SSEEdit
Backups' folder, so that's included as well. Also took this opportunity
to correct the 'SkyProc' directory (which should be 'SkyProc Patchers')
in the documentation.