Show a reason when ESL-checked plugin is already an ESL
Previously, we just displayed 'ESL Incapable' without any reason below
the file name when 'Check ESL Qualifications' was run on an ESL-flagged
file. This led to confusion for some users who weren't sure if this was
a bug or expected behavior. We now show 'Already ESL-flagged.' as a
reason, which should make this much more obvious.