Enable GMST merging for all games
Skyrim and its derivatives already had this enabled, as did CBash for
Oblivion. From some preliminary testing, there is no reason not to
enable this for all games - the record format is completely unchanged
since Oblivion.

Note that 'all games' does mean 'all games' here, including Fallout 4.
Shouldn't be a problem with the commit I pushed recently, but it does
make that commit a prerequisite for this one.