Slight reworking of mod list status icons RRR
Now shows a yellow one for plugin with our of order masters, since
that's not on the same level of problematic as a plugin loading before
its masters. Also refactored the ESL/ESM icon detection a bit and threw
in some better static text on hover.

Additionally, I finally got around to updating the status text/icon
section of the readme for ESLs, so threw that in here as well since it
touches the same section.

Related to # 309 <--- RRR, still need to add some sort of warning
message when activating broken plugins