Truncate and warn for overly long leveled lists RRR
Previously, we would just write these out, which struct.pack did not
like - duh, LLCT is an 8-bit counter:

ubyte format requires 0 <= number <= 255
<bash.brec.ModWriter instance at 0x0E1CDD00> LLCT B (264,)

Pretty cryptic, but what's worse is that it doesn't mention which
leveled list and which mod are responsible! This adds a new warning both
to the BashBugDump and the BP log, while also truncating lists to 255
entries to stop us writing out nonsense LVL* entries.

Took this opportunity to throw in a bunch of __repr__'s, both to make
BashBugDump errors nicer and for a better debugging experience when
stepping through records/patcher code.

Closes # 327 <--- RRR