Jump to Mods tab on double-click in Plugin Filter
A feature I wanted for a while, but totally forgot about until I
randomly stumbled upon 3d7b9816ec0e2b7d666a7bead0cd45db5347aed7
recently. This complements our 'Jump to Installer' feature in the Mods
tab by offering a way to jump right back to the mod that corresponds to
an installer. Also threw in a fixup for KeyErrors when double-clicking
empty space in the master list, mouse_index wasn't verified as
being >= 0.

Contains some readme edits to actually document the Plugin Filter and
some fixups for the version history where it was still called the 'Esp/m
Filter', as well as some noops to slightly improve _help and _text for
'Jump to Installer'.