Move all local absolute imports to relative
There were circular imports between bosh, parsers and record_groups that
did not happen with the absolute imports but had to be fixed with
relative imports. Everything else was pretty straightforward.

Infernio: Updated for 15, 461 and 480. Also moved all __author__s below
the imports to avoid syntax errors with __future__ imports.