Move building exe to pyinstaller
PyInstaller does not support pkgutil (or at least iter_modules) so the
call in bush.py was replaced by a simple list. There is another call in
localize.py but the function docs there say it's not meant to work when
frozen so I let it be. This is not really an ideal solution for it
should work for testing.

Translation scripts are in the wrong(?) location. I guess some scripts
expect them to be in Mopy/bash/ but not sure how that interacts with
PyInstaller - needs further testing.