Fix for WindowsError: [Error 6] The handle is invalid
In subprocess.py there is an if statement for when stderr is None.
This is the code that is failing because subprocess wants to access
`_subprocess.GetStdHandle(_subprocess.STD_ERROR_HANDLE)` and that errors
for some reason. The error in subprocess.py should be addressed however,
Python 2 is at end of life soon so I suggest this woraround when using
pyinstaller to build the EXE.