Been bothered with the menu
The items on the Menu apply to more then just a folder, they can apply to an archive as well. There is no reason to have "install Last" for example as something unique to a folder. It simply makes it install after the "==Last==" marker. You wouldn't need to make a folder to do that to a simple zip file.

I don't like the idea of needing all this submenus but if you want the context menu a bit shorter this is at least make more sense.