SSS Rework FOMOD GUI
Depends on inf-wx-begone, rewrites most of the GUI to use the wrappers
instead. Drops a whole bunch of wx usages, which is nice. RadioButton
needs wrapping, see all the ugly hacks at the bottom of gui_fomod. Also,
the design that uses __dict__ of wx objects to store group objects has
to go, it's fundamentally hacky and very fragile - e.g. imagine if the
wx guys decided to add __slots__ to their objects.

Also contains a bunch of fixes and misc improvements, e.g. user-facing
strings have been made translatable, some bugs that were carried over
from belt have been fixed, and the 'Back' button no longer works on the
first page.

Note the glaring TODOs - this is a straight up port of the original GUI,
but we currently don't have a way to change fonts, which the original
GUI relied on to differentiate its components. I added some
HBoxedLayouts as an alternative, which works fine for the main FOMOD
dialog and may even be an improvement in terms of visual clarity, but
doesn't help at all with the results screen, which is now an unreadable
mess.