env: Use KNOWNFOLDERID instead of CSIDL
As Microsoft's info page on these constants
(https://docs.microsoft.com/en-us/windows/win32/shell/csidl) mentions:

 As of Windows Vista, these values have been replaced by KNOWNFOLDERID
 values. [...] The CSIDL system is supported under Windows Vista for
 compatibility reasons. However, new development should use
 KNOWNFOLDERID values rather than CSIDL values.

I suspect this is the cause of the IOError: [Errno 22] Invalid argument
that some users are getting - for one user, the CSIDL values resolved to
C:\Users\{USER NAME HERE}\OneDrive\My Documents\, which is obviously
wrong. In either case, we should be using the new ones here.

For some reason, pywin32 doesn't provide wrappers for these. So this is
based on code by Michael Kropat, see the code in env.py for full credits
and a link to the original. I don't like that this introduces some
ctypes-based code to env.py, I'd rather have that in windows.py, but a
circular dependency would arise if I tried that...