Remove incorrect warning from bash_default.ini
Introduced in 368b8509c0800010de1cb669fca92783a87417b3, due to this bug
report on sourceforge: https://sourceforge.net/p/oblivionworks/bugs/224/

Then edited in 83409c881b4ffc8439d479b34136c4bf5aae3eee to remove the
'or Skyrim' part, leaving us with an entirely incorrect warning.

The original intent was to warn people not to use this option on Skyrim
because it would cause Steam Workshop mods to constantly redownload
(they use the timestamp of files to determine if redownloading is
necessary). But then the option was completely disabled for Skyrim
because the new BSA system in that game obsoletes it. Oblivion never had
a Steam Workshop, so it should not have been mentioned in there at all.